// nhl-stats-app-client/src/types/constants.ts
export const PLAYER_STATS_URL =
  'https://www.rotowire.com/hockey/tables/stats.php?pos=skater&season=2024';

export const PAGE_SIZE = 10;

export enum StatCategories {
  // PLAYER_NAME = 'Player',
  TEAM = 'team',
  POS = 'pos',
  GAMES_PLAYED = 'gp',
  GOALS = 'goals',
  ASSISTS = 'assists',
  POINTS = 'points',
  SHOTS_ON_GOAL = 'shots',
  HITS = 'hits',
  BLOCKS = 'blocks',
}